Reduced functions in config into individual files.
