--- src/scripts/dimApp.hotkeys.config.js
@@ -1,5 +1,0 @@
-export function hotkeys(hotkeysProvider) {
-  "ngInject";
-
-  hotkeysProvider.includeCheatSheet = true;
-}
