--- src/scripts/dimApp.http.config.js
@@ -1,9 +1,0 @@
-export function http($httpProvider) {
-  "ngInject";
-
-  $httpProvider.interceptors.push("ngHttpRateLimiterInterceptor");
-
-  if (!window.chrome || !window.chrome.extension) {
-    $httpProvider.interceptors.push('http-refresh-token');
-  }
-}
