--- src/scripts/dimApp.local-storage.config.js
@@ -1,5 +1,0 @@
-export function localstorage(localStorageServiceProvider) {
-  "ngInject";
-
-  localStorageServiceProvider.setPrefix('');
-}
