+++ src/scripts/services/dimLoadingTracker.factory.js
@@ -1,0 +1,7 @@
+function loadingTracker(promiseTracker) {
+  "ngInject";
+
+  return promiseTracker();
+}
+
+export default loadingTracker;
