+++ src/scripts/state.js
@@ -1,0 +1,7 @@
+const state = {
+  membershipType: -1,
+  active: null,
+  debug: false
+};
+
+export default state;
